chore: Move comments event handler to use proper event dispatcher
