+++ lib/Listeners/CommentEventListener.php
@@ -1,43 +1,37 @@
 <?php
+
+declare(strict_types=1);
 /**
  * SPDX-FileCopyrightText: 2018 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-namespace OCA\Deck\Activity;
+namespace OCA\Deck\Listeners;
 
+use OCA\Deck\Activity\ActivityManager;
 use OCA\Deck\Db\CardMapper;
 use OCA\Deck\Db\ChangeHelper;
 use OCA\Deck\Notification\NotificationHelper;
 use OCP\Comments\CommentsEvent;
-use OCP\Comments\IComment;
-use OCP\Comments\ICommentsEventHandler;
+use OCP\EventDispatcher\Event;
+use OCP\EventDispatcher\IEventListener;
 
-class CommentEventHandler implements ICommentsEventHandler {
+/** @template-implements IEventListener<CommentsEvent|Event> */
+class CommentEventListener implements IEventListener {
 
-	/** @var ActivityManager */
-	private $activityManager;
-
-	/** @var NotificationHelper */
-	private $notificationHelper;
-
-	/** @var CardMapper */
-	private $cardMapper;
-
-	/** @var ChangeHelper */
-	private $changeHelper;
-
-	public function __construct(ActivityManager $activityManager, NotificationHelper $notificationHelper, CardMapper $cardMapper, ChangeHelper $changeHelper) {
-		$this->notificationHelper = $notificationHelper;
-		$this->activityManager = $activityManager;
-		$this->cardMapper = $cardMapper;
-		$this->changeHelper = $changeHelper;
+	public function __construct(
+		private ActivityManager $activityManager,
+		private NotificationHelper $notificationHelper,
+		private CardMapper $cardMapper,
+		private ChangeHelper $changeHelper,
+	) {
 	}
 
-	/**
-	 * @param CommentsEvent $event
-	 */
-	public function handle(CommentsEvent $event) {
+	public function handle(Event $event): void {
+		if (!$event instanceof CommentsEvent) {
+			return;
+		}
+
 		if ($event->getComment()->getObjectType() !== 'deckCard') {
 			return;
 		}
@@ -61,20 +55,13 @@ class CommentEventHandler implements ICommentsEventHandler {
 		}
 	}
 
-	/**
-	 * @param CommentsEvent $event
-	 */
-	private function activityHandler(CommentsEvent $event) {
-		/** @var IComment $comment */
+	private function activityHandler(CommentsEvent $event): void {
 		$comment = $event->getComment();
 		$card = $this->cardMapper->find($comment->getObjectId());
 		$this->activityManager->triggerEvent(ActivityManager::DECK_OBJECT_CARD, $card, ActivityManager::SUBJECT_CARD_COMMENT_CREATE, ['comment' => $comment]);
 	}
 
-	/**
-	 * @param CommentsEvent $event
-	 */
-	private function notificationHandler(CommentsEvent $event) {
+	private function notificationHandler(CommentsEvent $event): void {
 		$this->notificationHelper->sendMention($event->getComment());
 	}
 }
